refactor: file structure in root
