+++ src/lib/components/reusable/StepperConv.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-    import {exBranches, exCommits, exIssues} from "../../enums/enums";
+    import {exBranches, exCommits, exIssues} from "../../../typescript/enums/enums";
 
     export let type:string;
     export let example:number;
