+++ src/typescript/classes/Convention.ts
