+++ src/typescript/classes/Payload.ts
